move API schemas to types/api dir
